Use the higher of `Suggested` and `Minimum` strictness when autocorrecting
We want to use config like this:

```yml
Sorbet/FalseSigil:
  Enabled: true
  SuggestedStrictness: true
```

So that we can require at least a `false` sigil on *all* files, but add `typed: true` to new files. This way we know that Sorbet adoption will go up over time as new files are added (we have a separate Danger rule that warns against downgrading a file's type).

This PR adds the ability to do that, by choosing the higher level out of the suggested and minimum levels when autocorrecting. There is a special case when you are using `ignore` as your minimum level in which this doesn't take place, I did this rather than change existing specs so as to not create a breaking change.