Fix ParametersOrderingInSignature crash on invalid sig
`ParametersOrderingInSignature` was erroring out when it was not
supplied a parameter set that was a `HashNode` since the code was
assuming it would always get a valid signature.

We now check for the node type before accessing its keys and add an
invalid signature offense if the `params` block is malformed.