Preserve generic formatting by using index send nodes
Fixes: #40

By default RuboCop does not enable the more modern AST syntax that users
of whitequark parser can opt-in to by turning on various `emit_xxx`
flags on `Parser::Builders::Default`. One of those modern Ruby AST
syntax is the `index` nodes that can be emitted for index accessors,
which is necessary for us to properly roundtrip generic types back to
expected Ruby code.

The "fix" is an ugly hack that re-parses the given node source with
`emit_index` turned on and applies `unparse` on the returned node set
instead of the original one.