Fix default config for `RedundantExtendTSig`
In the original draft, the cop was called `ForbidExtendTSig`. It was
later renamed, and the config was moved to what would be the correct
location alphabetically, but the actual config key was not updated.

This cause the cop to have no default config, so it was enabled by
default, which would have been problematic for consumers.