Add `Sorbet/TStructPropertyAttrMacro`
In a `T::Struct`, `attr_*` methods shouldn't be used for properties.
Instead, `prop` should be used instead of `const` to make a mutable
property, and all simple accesses should go through the generated
methods.
The only case where a custom reader or writer should be used is if the
logic is being customized.