Optimize `EmptyLineAfterSig`
The previous implementation would traverse the tokens in the file each
time it would analyze a `sig` until it found the corresponding method
definition. This would take additional time the further we got into a
file, as we would have to re-traverse all tokens so far.

This new implementation simply looks for the next node after the `sig`,
confirms it is the method definition, and calculates offsets for the
lines in between.