Improve `CallbackConditionalsBinding` specs
Specs are consolidated by using `expect_correction` to test offenses and
correction in the same example.

Doing this reveals that we actually have corrections which do not handle
whitespace correctly. For now, the incorrect whitespace has been
included in the specs, so they reflect the existing behavior.

In addition, some examples are added covering edge cases, to capture the
existing behavior ahead of changing it.