Add task to release this gem
I noticed that some cops were with the `<<next>>` version in the config
file. This was supposed to be changed when a new release was made,
but it is too easy to forget.

Instead, we should just codify this in a task so anyone that needs to
prepare a release don't need to remember what they are supposed to do.